Move code to point of use
